# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A Go service for investigating \"Deep inspection\" of VMs in VMware vSphere",
        "title": "VM Deep Inspection Demo API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/inspector/v2v": {
            "post": {
                "description": "Queue V2V-only inspection jobs through the selected VDDK or NFC backend. Each job creates and removes a temporary vSphere snapshot.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inspector"
                ],
                "summary": "Start asynchronous virt-v2v inspection",
                "parameters": [
                    {
                        "description": "VM names to inspect",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.startV2VRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Jobs accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "A VM already has an active job or the queue is full",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Selected inspection backend is unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inspector"
                ],
                "summary": "Cancel all queued and running V2V inspections",
                "responses": {
                    "202": {
                        "description": "Cancellation requested",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/inspector/v2v/status": {
            "get": {
                "description": "Return the latest persisted V2V job state for each VM.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inspector"
                ],
                "summary": "Get V2V inspection status",
                "responses": {
                    "200": {
                        "description": "V2V status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Status query failed",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/inspector/v2v/{vmName}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inspector"
                ],
                "summary": "Cancel V2V inspection for one VM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM name",
                        "name": "vmName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Cancellation requested",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "No active job",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms": {
            "get": {
                "description": "Get a list of all virtual machines with optional name filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "List all virtual machines",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web\"",
                        "description": "Filter VMs where name contains this string",
                        "name": "name_contains",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of virtual machines",
                        "schema": {
                            "$ref": "#/definitions/types.VMListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "vSphere connection unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/check": {
            "post": {
                "description": "Run validation checks on a VM snapshot. If check parameter is provided, runs that specific check. If omitted, runs all available checks.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Run validation checks on a VM snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "Original VM name",
                        "name": "vm",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"inspection-snapshot\"",
                        "description": "Snapshot name",
                        "name": "snapshot",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"fstab\"",
                        "description": "Check type to run (fstab, disk-access, bsod). If omitted, runs all checks.",
                        "name": "check",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check completed successfully",
                        "schema": {
                            "$ref": "#/definitions/types.CheckResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM or snapshot not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/clone": {
            "post": {
                "description": "Create a linked clone from a VM snapshot for inspection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Create a clone from VM snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "VM name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Clone request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CloneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clone created successfully",
                        "schema": {
                            "$ref": "#/definitions/types.CloneResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM or snapshot not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/delete-clone": {
            "delete": {
                "description": "Delete a cloned VM created for inspection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Delete a cloned VM",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01-clone-123\"",
                        "description": "Clone VM name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clone deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Clone not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/inspect-snapshot": {
            "post": {
                "description": "Run virt-inspector on a VM snapshot through the selected inspection backend. Optionally include virt-v2v-inspector results.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Inspect a VM snapshot directly",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "Original VM name",
                        "name": "vm",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"inspection-snapshot\"",
                        "description": "Snapshot name",
                        "name": "snapshot",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"virt-inspector\"",
                        "description": "Inspector type: 'virt-inspector' (default, runs only virt-inspector) or 'virt-v2v-inspector' (runs both virt-inspector and virt-v2v-inspector)",
                        "name": "inspector",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inspection completed successfully",
                        "schema": {
                            "$ref": "#/definitions/types.VMInspectionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM or snapshot not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/snapshot": {
            "post": {
                "description": "Create a snapshot for a specific virtual machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Create a VM snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "VM name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Snapshot creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SnapshotCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Snapshot created successfully",
                        "schema": {
                            "$ref": "#/definitions/types.SnapshotCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "vSphere connection unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{name}": {
            "get": {
                "description": "Get detailed information about a specific virtual machine by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Get virtual machine details",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "VM name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Virtual machine details",
                        "schema": {
                            "$ref": "#/definitions/types.VMDetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "vSphere connection unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.startV2VRequest": {
            "type": "object",
            "required": [
                "vm_names"
            ],
            "properties": {
                "vm_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "types.CheckResponse": {
            "type": "object",
            "properties": {
                "all_valid": {
                    "type": "boolean",
                    "example": true
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CheckResult"
                    }
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "backup-snapshot"
                },
                "vm_name": {
                    "type": "string",
                    "example": "web-server-01"
                }
            }
        },
...
        "types.CloneRequest": {
            "type": "object",
            "required": [
                "snapshot_name"
            ],
            "properties": {
                "clone_name": {
                    "type": "string",
                    "example": "my-clone"
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "backup-snapshot"
                }
            }
        },
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "string",
                    "example": "VM ID is required"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
        "types.VM": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "web-server-01"
                },
                "power_state": {
                    "type": "string",
                    "example": "poweredOn"
                },
                "uuid": {
                    "type": "string",
                    "example": "502e7c6e-b5c3-4d0e-9a5a-8b9c1d2e3f4g"
                }
            }
        },
...
        "types.VMListResponse": {
            "type": "object",
            "properties": {
                "datacenter": {
                    "type": "string",
                    "example": "Datacenter1"
                },
                "total": {
                    "type": "integer",
                    "example": 150
                },
                "vms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.VM"
                    }
                }
            }
        },
...
    }
}